modificato la struttura del progetto. Aggiunto controllo dopo l'esecuzione di ogni query all'interno del file.cpp. Creato il file di creazione e popolamento delle tabelle che dovra' poi essere consegnato.
