+++ queries/entity.sql
